Separate services in main.ts into single files
