+++ src/styles/sass/components/_image.scss
@@ -56,6 +56,8 @@
     img {
       width: 100%;
       height: auto;
+      max-height: 400px;
+      object-fit: cover;
     }
   }
 }
