+++ types/services/section.service.d.ts
@@ -1,0 +1,14 @@
+import { BlockService } from './block.service';
+export declare class SectionService {
+    private blockService;
+    constructor(blockService: BlockService);
+    init(container: HTMLElement): void;
+    /**
+     * Decorates all sections in a container element.
+     * @param {Element} main The container element
+     */
+    private transformSection;
+    private processSectionMetaData;
+    private adjustMarkup;
+    decorateImages(): void;
+}
